requirements.txt
flask
gunicorn
requests
colorama
openai>=1.12.0
python-dotenv
